+++ src/main/java/frc/robot/subsystems/Auton/AutoDirector.java
@@ -4,14 +4,15 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Supplier;
 
+import com.ctre.phoenix6.swerve.SwerveRequest;
 import com.pathplanner.lib.auto.AutoBuilder;
 
 import edu.wpi.first.math.geometry.Pose2d;
-import edu.wpi.first.networktables.NetworkTable;
-import edu.wpi.first.networktables.NetworkTableInstance;
 import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
+import edu.wpi.first.wpilibj2.command.Commands;
+import edu.wpi.first.wpilibj2.command.ConditionalCommand;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import frc.robot.Constants.AutoConstants.AutoSector;
 import frc.robot.Constants.AutoConstants.AutoSectorV2;
@@ -51,14 +52,17 @@ public class AutoDirector {
     // Autos.add(PathPlannerAuto());
     // Autos.add(HPintake());
     // Autos.add(newAuto());
-    Autos.add(Leftside());
+    // Autos.add(Leftside());
+    // Autos.add(RightSide());
+    // Autos.add(CenterSide());
+    // Autos.add(CenterSide3());
+    // Autos.add(CenterSideV2());
+    // Autos.add(CenterSideV3());
+    // Autos.add(testHP());
+    // Autos.add(dummyAuto());
+    Autos.add(LeftSide());
     Autos.add(RightSide());
-    Autos.add(CenterSide());
-    Autos.add(CenterSide3());
-    Autos.add(CenterSideV2());
-    Autos.add(CenterSideV3());
-    Autos.add(testHP());
-    Autos.add(dummyAuto());
+    Autos.add(Middle());
     setupSmartAuto();
     for (Auto auto : Autos) {
       autoChooser.addOption(auto.name, auto);
@@ -67,10 +71,17 @@ public class AutoDirector {
   }
 
   // ------------------------------------------Autos------------------------------------------
+  // public Auto doNothing() {
+  //   return new Auto("doNothing", new InstantCommand(), new Pose2d());
+  // }
   public Auto doNothing() {
-    return new Auto("doNothing", new InstantCommand(), new Pose2d());
-  }
+    Command moveXPos = (subsystems.driveSubsystem().Commands.applyRequest(() ->  new SwerveRequest.RobotCentric().withVelocityX(1)));
+    Command speedZero = subsystems.driveSubsystem().Commands.applyRequest(() -> new SwerveRequest.RobotCentric().withVelocityX(0));
+    Command cmd = moveXPos.withDeadline(Commands.waitSeconds(1)).andThen(speedZero);
 
+    // var cmd = subsystems.driveSubsystem().setControl(subsystems.driveSubsystem().Commands.applyRequest(() ->  new SwerveRequest.RobotCentric().withVelocityX(1)));
+    return new Auto("doNothing", cmd, new Pose2d());
+  }
   public Auto newAuto() {
     List<AutoSector> paths = new ArrayList<>();
     paths.add(new AutoSector("TSC", "R6B"));
@@ -128,16 +139,7 @@ public class AutoDirector {
     return new Auto("Leftside 4 Coral", tracker, PositionConstants.startingPoses.Left());
   }
 
-  public Auto RightSide() {
-    List<AutoSector> paths = new ArrayList<>();
-    paths.add(new AutoSector("RHP", "R2A"));
-    paths.add(new AutoSector("RHP", "R2B"));
-    paths.add(new AutoSector("RHP", "R4A"));
-
-    AutoTracker tracker = new AutoTracker(subsystems, paths, () -> PositionConstants.startingPoses.right(), "R12B");
-
-    return new Auto("Rightside 4 Coral", tracker, PositionConstants.startingPoses.right());
-  }
+  
 
   public Auto CenterSide() {
     List<AutoSector> paths = new ArrayList<>();
@@ -183,6 +185,33 @@ public class AutoDirector {
     AutoTrackerV2 tracker = new AutoTrackerV2(subsystems, () -> PositionConstants.startingPoses.CenterLeft());
     tracker.addSector(new AutoSectorV2(HumanPlayerIntake.BlueLeft, Reef.BlueR8A));
     return new Auto("Testing HP intake", tracker);
+  }
+
+  public Auto LeftSide(){
+    AutoTrackerV2 tracker = new AutoTrackerV2(subsystems, ()-> PositionConstants.startingPoses.RightMidBarge());
+    tracker.addPoint(new Pose2d(1.232, 3.071, PositionConstants.awayFromAlliance));
+    tracker.addPreload(Reef.BlueR4A);
+    tracker.addSector(new AutoSectorV2(GamePeice.BlueC3, Reef.BlueR4B));
+    tracker.addSector(new AutoSectorV2(GamePeice.BlueC2, Reef.BlueR6A));
+
+    return new Auto("LeftSide", tracker);
+  }
+
+  public Auto RightSide(){
+    AutoTrackerV2 tracker = new AutoTrackerV2(subsystems, ()-> PositionConstants.startingPoses.LeftBargeMiddle());
+    tracker.addPoint(new Pose2d(1.848, 4.934, PositionConstants.awayFromAlliance));
+    tracker.addPreload(Reef.BlueR6B);
+    tracker.addSector(new AutoSectorV2(GamePeice.BlueC2, Reef.BlueR6A));
+    tracker.addSector(new AutoSectorV2(GamePeice.BlueC1, Reef.BlueR8A));
+
+    return new Auto("RightSide", tracker);
+  }
+
+  public Auto Middle(){
+    AutoTrackerV2 tracker = new AutoTrackerV2(subsystems, ()-> PositionConstants.startingPoses.Center());
+    tracker.addPreload(Reef.BlueR12B);
+
+    return new Auto("Middle", tracker);
   }
 
   public Auto auto(String name, List<AutoSector> paths, Supplier<Pose2d> initalPose,