+++ .github/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0]
+### Changed
+- Arm Gear Ratio for new Chain
+- Added Comp Autos
+- Coral Cam Code Adjustment
+- Adjusted Constants for positions for the comp field
+
 ## [0.5.0]
 ### Changed
 - Dead Zones